Better nextBox logic (inside putInBox function). Using items instead of value for end of packing logic
3.0.15  2021-05-06 ---
Added itemsTotalLeft (total number of items left to pack) to replace totalValueCurrent (total space of items) in some places, since some items are worth 0 so it was acting like it was done packing with SSR left.

Added missing valuePam1 into valueTotal and totalValueTotal.

Removed if case paramaters for when to calculate putting items into a large accessory box (now does it after every box is finished).

Added an else case after a box is finished, to reset back to accessory box (was staying on large accessory box if that was previous).

Moved a lot of nextBox code from each item into the putInBox function after finishing a box, so the logic is a lot better for picking the next box now.
(Left nextBox code in the insulPouch section since it needs to force itself into a separate medium shipper every time)

TODO: Fix something not getting subtracted from itemsTotalLeft

---
After packing a 14x14, do another if statement to see if we should stay on 14x14 or switch back to an accessory box (was packing 2 fourHour timers in a 2nd 14x14).

Fix/stop showing weight after leak rope / before ssr. Problem was it shows weight if total value of current items to pack left is 0, and ssr are worth 0.
Changing to compare total number of items left instead of total calculated item space left to pack.
---